+++ src/lib.rs
@@ -19,8 +19,10 @@
 
 extern crate bare_metal;
 extern crate bit_field;
+extern crate embedded_hal;
 
 pub mod asm;
+pub mod delay;
 pub mod interrupt;
 pub mod register;