+++ src/register/marchid.rs
@@ -10,6 +10,7 @@ pub struct Marchid {
 
 impl Marchid {
     /// Returns the contents of the register as raw bits
+    #[inline]
     pub fn bits(&self) -> usize {
         self.bits.get()
     }