+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - add enums for PMP configuration options (Range, Permission)
 - add set_pmp() and clear_pmp() functions to pmpcfg(x) modules 
 
+### Changed
+
+- Use new `asm!` instead of `llvm_asm!`
+
 ## [v0.7.0] - 2020-07-29
 
 ### Added