+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Add enums `Range`, `Permission` for PMP configuration
 - Add `set_pmp()` and `clear_pmp()` functions to pmpcfg(x) modules
 - Add struct `Pmpcsr` and is returned from `pmpcfgx::read()`
-- Add delay structure and methods using embedded-hal traits and Mcycle register
+- Add delay structure and methods using embedded-hal traits and `mcycle` register
 - add asm::delay function for assembly-based busy-loops
 - add asm::nop, a wrapper for implementing a `nop` instruction