+++ CHANGELOG.md
@@ -6,9 +6,36 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Add `#[cfg(riscv32)]` to `pmpcfg1` and `pmpcfg3` modules
+- Add enums `Range`, `Permission` for PMP configuration
+- Add `set_pmp()` and `clear_pmp()` functions to pmpcfg(x) modules
+- Add struct `Pmpcsr` and is returned from `pmpcfgx::read()`
+- Add delay structure and methods using embedded-hal traits and `mcycle` register
+- Add `asm::delay()` function for assembly-based busy-loops
+- Add `asm::nop()`, a wrapper for implementing a `nop` instruction
+- Add missing `#[inline]` attribute to register reads, type conversations and `interrupt::free`
 
 ### Changed
 
+- Use new `asm!` instead of `llvm_asm!`
+- Change `pmpcfgx::read()` macro to `read_csr_as!()` from `read_csr_as_usize!()` 
+
+## [v0.7.0] - 2020-07-29
+
+### Added
+
+- Add `medeleg` register
+- Add `cycle[h]`, `instret[h]` and `mcounteren`
+- Add additional binaries for floating-point ABIs
+- Add support for `mxr`
+- Add support for `mprv`
+
+### Changed
+
+- Fix `scause::set`
+- Various formatting and comment fixes
+- Update `bare-metal` to `v1.0.0` removing `Nr` trait
 - Build targets on `docs.rs` are now RISC-V targets other than default ones
 
 ## [v0.6.0] - 2020-06-20